task 3 finished for all the provided cases
created array of total scores for each player and compared each to the dealer's score. Accounted for cases where the player has less than or equal to the 21 and at the same time has more than or equal to the dealer. Also the cases where the dealer busts and the player does not.